api/requirements.txt
Django==1.8.2
South==0.8.1
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
djangorestframework==3.1.0
gunicorn==19.3.0
psycopg2==2.6
selenium==2.46.0
solid-i18n==1.1.0
static3==0.6.1
wsgiref==0.1.2
